added battery parse on reset button press
After the device is properly paired with SmartThings, a short press of the reset / pairing button results in the sensor sending a non-standard announcement hexadecimal message string which among other things contains the current battery voltage value.

Pressing the button to generate the announcement string does not have any effect on the normal checkin interval and also does not prevent the initial or any subsequent catchall messages that also provide battery voltage data.

Like the open / close and catchall events, the button press announcement event is picked up for parsing via zigbee.getEvent, so to handle that, a new condition is added to def parse(String description) that acts on a message string starting with "read attr - raw:".

Because the button-press message string data is non-standard, it cannot be parsed using any available zigbee parse function, and also data.get cannot be used to capture specific bytes from the string. Instead it must be handled as text, and after checking for the presence of the attrId 01FF, the two bytes of the battery voltage data are converted to an integer which is passed on to the already existing getBatteryResult function.