Code overhaul - with continuous motion detection!!
I have gone through all of the Aqara Motion Detector DTH code, and after removing unnecessary function calls and lines of code, removing error-prone code, and looking at the motion reset preference issue, I have discovered that this device is capable of continuous motion detection!!! This is an amazing revelation considering it's price.

It turns out that as long as there is motion detected, it repeatedly sends a read attr - raw: message, usually along with an illuminance reading. The message is the same every time: endpoint: 01, cluster: 0406, size: 08, attrId: 0000, encoding: 18, value: 01. As soon as the motion stops, the sensor stops sending the message, but it doesn't send any "inactivity" message (which would be value 00 in this case).

So the trick is to still use the runIn() scheduler to reset back to the 'no motion' / inactivity state that was already in the code, but find the magic number of seconds that will allow the motion detection events to overlap and is short enough to go back to no motion very soon after the motion has actually stopped being detected.

The magic number I came up with in my testing is 7 seconds. Shorter than that and then the state alternates between motion & no motion while I move around. Longer than 7, and it starts to feel like a lag or delay after I've stopped moving.

However, I assume other people may find a different ideal number, so I changed the preference setting slightly, to only accept values from 1 to 60 seconds, and the description has been changed (and it's much shorter!)

I am extremely happy with how the code is working, but I really want some others to test it to see how if works for them.

I also need to go through the "original" motion detector code to clean it up as well, and then someone who actually has one of those needs to do some testing to find out if it also is capable of continuous motion detection.

Note that I created this Pull Request commit off of @tmleafs Standardization branch, so it should be merged first and this one afterwards.