Added a4refillpad method to determine button "held"
I have reincorporated a4refillpad's method to determine whether a button "pushed" or "held" event is sent, based on the difference in time between pressing the button and releasing it.

By all reports, this method seems to be the most reliable, but still not 100% because of some press & release messages being parsed in reversed order, or even dropped by the hub. However, it's main limitation is that after pressing the button, nothing will happen until you release it. So even if the `waittoHeld` preference is set to 2 seconds, if the button is held for 10 seconds, the button "held" event only gets sent at 10 seconds, when the button is physically released.

Note that I've re-worked a4refillpad's code to remove redundancies, and to make sure the main `parse(String description)` routine is the only one calling `createEvent()`.

To try to make this DTH more versatile for WebCoRE users, I've also added `lastReleased` and `lastReleasedDate` events that are triggered when the button is physically released.

I haven't tested this DTH yet, and it really must be tested before it's released.

If this code works well enough, then I can use it to build the DTH for the new Aqara button model.