Use a4refillpad method to determine button "held"
Fixes
• Added missing `if {}` condition brackets
• Added `lastButtonMssg` attribute to send events any time the button is pressed or released
• changed the time difference calculation to compare time of last `lastButtonMssg` event to current time at button release
• changed default button held time preference setting to 2 seconds

Button pressed vs held seems to be around 95% accurate and reliable, with a `waittoHeld` setting of 2 seconds or longer. This is as good as it's going to get as far as I can tell, because the ST hub still drops some button press/release messages and also sometimes gets them in reverse.