More fixes - no errors now
I discovered the reason for the error on pairing, and modified the code that handles `attr read - raw:` messages on `cluster: 0000` / `attrId: 0005` so that the short announcement message with just the model name is parsed correctly in addition to the longer announcement message with model name and battery voltage info.

Also, I made some discoveries about how to improve the code that sets things up during the initial `installed()`, `configure()`, and `updated()` calls just after pairing. These improvements will need to be applied to all other DTHs, because there were some things that weren't working correctly.