Added fingerprints for models WXKG01LM / WXKG02LM
I discovered I had forgot to update the name of a function call used for the original Aqara Button (model WXKG11LM). So that's fixed.

Then I added fingerprint info for both of the Aqara Smart Light Switch models (WXKG01LM and WXKG02LM). Note that for the 2-button model, SmartThings treats it as a SINGLE button, because the Zigbee API strips out the `endpoint` number from the button press message, so all three types of button presses (left, right, both) are received by the DTH as `on/off: 1`, unfortunately. There is no way to fix this, because it requires SmartThings to change its API.

I also fixed an error that could happen if someone pairs a button using the manual "catchall" method.